MB-67520: Fix narrowing warnings in ep-benchmarks
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231030
Tested-by: Jim Walker <jim@couchbase.com>

+++ engines/ep/benchmarks/hash_table_bench.cc
@@ -47,7 +47,7 @@ public:
 
     StoredDocKey makeKey(CollectionID collection,
                          std::string_view keyPrefix,
-                         int i) {
+                         size_t i) {
         // Use fmtlib to format key with stack-local (non-heap) buffer to
         // minimise the cost of constructing keys for Items.
         fmt::memory_buffer keyBuf;
